优化缩进代码块的逻辑，兼容高亮代码块和行内代码块互相嵌套的场景
