+++ src/Cherry.config.js
@@ -70,21 +70,6 @@ const defaultConfig = {
       // 'hookName': {
       //
       // }
-      pre: {
-        /**
-         * 是否启用四个空格生成代码块的语法
-         *
-         * 为什么增加这个开关？
-         * 答：
-         *    在6.X之前的版本中默认不支持该语法。
-         *    因为cherry的开发团队认为该语法太丑了（容易误触）
-         *    开发团队希望用```代码块语法来彻底取代该语法
-         *    但在后续的沟通中，开发团队发现在某些场景下该语法有更好的显示效果
-         *    因此开发团队在6.X版本中才引入了该语法
-         *    为了让已经引用cherry的业务无感知，我们提供了关闭该语法的配置
-         */
-        enable: true,
-      },
       list: {
         listNested: false, // 同级列表类型转换后变为子级
         indentSpace: 2, // 默认2个空格缩进
@@ -104,6 +89,18 @@ const defaultConfig = {
         customRenderer: {
           // 自定义语法渲染器
         },
+        /**
+         * indentedCodeBlock是缩进代码块是否启用的开关
+         *
+         *    在6.X之前的版本中默认不支持该语法。
+         *    因为cherry的开发团队认为该语法太丑了（容易误触）
+         *    开发团队希望用```代码块语法来彻底取代该语法
+         *    但在后续的沟通中，开发团队发现在某些场景下该语法有更好的显示效果
+         *    因此开发团队在6.X版本中才引入了该语法
+         *    已经引用6.x以下版本的业务如果想做到用户无感知升级，可以去掉该语法：
+         *        indentedCodeBlock：false
+         */
+        indentedCodeBlock: true,
       },
       emoji: {
         useUnicode: true, // 是否使用unicode进行渲染
